+++ src/client/request/graphql/generators/credit-card-tokenization.js
@@ -2,31 +2,49 @@
 
 var assign = require('../../../../lib/assign').assign;
 
-var CREDIT_CARD_TOKENIZATION_MUTATION = 'mutation TokenizeCreditCard($input: TokenizeCreditCardInput!) { ' +
-'  tokenizeCreditCard(input: $input) { ' +
-'    token ' +
-'    creditCard { ' +
-'      bin ' +
-'      brandCode ' +
-'      last4 ' +
-'      expirationMonth' +
-'      expirationYear' +
-'      binData { ' +
-'        prepaid ' +
-'        healthcare ' +
-'        debit ' +
-'        durbinRegulated ' +
-'        commercial ' +
-'        payroll ' +
-'        issuingBank ' +
-'        countryOfIssuance ' +
-'        productId ' +
-'      } ' +
-'    } ' +
-'  } ' +
-'}';
+function createMutation(config) {
+  var hasAuthenticationInsight = config.hasAuthenticationInsight;
+  var mutation = 'mutation TokenizeCreditCard($input: TokenizeCreditCardInput!';
 
-function createCreditCardTokenizationBody(body) {
+  if (hasAuthenticationInsight) {
+    mutation += ', $authenticationInsightInput: AuthenticationInsightInput!';
+  }
+
+  mutation += ') { ' +
+    '  tokenizeCreditCard(input: $input) { ' +
+    '    token ' +
+    '    creditCard { ' +
+    '      bin ' +
+    '      brandCode ' +
+    '      last4 ' +
+    '      expirationMonth' +
+    '      expirationYear' +
+    '      binData { ' +
+    '        prepaid ' +
+    '        healthcare ' +
+    '        debit ' +
+    '        durbinRegulated ' +
+    '        commercial ' +
+    '        payroll ' +
+    '        issuingBank ' +
+    '        countryOfIssuance ' +
+    '        productId ' +
+    '      } ' +
+    '    } ';
+
+  if (hasAuthenticationInsight) {
+    mutation += '    authenticationInsight(input: $authenticationInsightInput) {' +
+      '      customerAuthenticationRegulationEnvironment' +
+      '    }';
+  }
+
+  mutation += '  } ' +
+    '}';
+
+  return mutation;
+}
+
+function createCreditCardTokenizationBody(body, options) {
   var cc = body.creditCard;
   var billingAddress = cc && cc.billingAddress;
   var expDate = cc && cc.expirationDate;
@@ -44,6 +62,12 @@ function createCreditCardTokenizationBody(body) {
       options: {}
     }
   };
+
+  if (options.hasAuthenticationInsight) {
+    variables.authenticationInsightInput = {
+      merchantAccountId: body.merchantAccountId
+    };
+  }
 
   if (billingAddress) {
     variables.input.creditCard.billingAddress = billingAddress;
@@ -75,9 +99,13 @@ function addValidationRule(body, input) {
 }
 
 function creditCardTokenization(body) {
+  var options = {
+    hasAuthenticationInsight: Boolean(body.authenticationInsight && body.merchantAccountId)
+  };
+
   return {
-    query: CREDIT_CARD_TOKENIZATION_MUTATION,
-    variables: createCreditCardTokenizationBody(body),
+    query: createMutation(options),
+    variables: createCreditCardTokenizationBody(body, options),
     operationName: 'TokenizeCreditCard'
   };
 }