+++ src/three-d-secure/shared/constants.js
@@ -3,7 +3,7 @@
 module.exports = {
   LANDING_FRAME_NAME: 'braintreethreedsecurelanding',
   CARDINAL_SCRIPT_SOURCE: {
-    production: 'https://songbird.cardinalcommerce.com/cardinalcruise/v1/songbird.js',
-    sandbox: 'https://songbirdstag.cardinalcommerce.com/cardinalcruise/v1/songbird.js'
+    production: 'https://songbird.cardinalcommerce.com/edge/v1/songbird.js',
+    sandbox: 'https://songbirdstag.cardinalcommerce.com/edge/v1/songbird.js'
   }
 };